Update Fun.js
Funbot.misc.weedfact = [
"No one has ever overdosed or died from smoking too much weed. In the history of mankind.",
"Marijuana has been in an “official” state of prohibition since 1937-s Marijuana Tax Act (in the United States). Because, among other reasons, it made white women desire black men.",
"Colorado and Washington have estimated tax revenues of over $550 Million from legalization",
"There are more medical marijuana dispensaries than Starbucks in Los Angeles and Denver. Because reefer is safer than coffee.",
"The first law in America relating to cannabis was a 1619 law that actually required farmers to grow the hemp plant. Once harvested, hemp was useful for clothing, sails, and rope.",
"In 1906, cannabis was labeled as a “poison” and states began restricting its sale.",
"By the mid 1930s, cannabis was regulated as a drug in every state.",
"From 1850 to 1942, marijuana was listed in the United States Pharmacopoeia as a useful medicine for nausea, rheumatism, and labor pains and could be purchased from pharmacies and general stores.",
"The last four presidents of the U.S, including Barack Obama, have all admitted to smoking cannabis.",
"Since Nixon declared the “War On Drugs” in 1971, over a trillion dollars have been spent on it. It takes the average person under an hour to find drugs in a new city.",
"A recent Gallup Poll concluded that 52% of Americans now favor all out legalization. The other 48% are packing their bongs. And 72% of Americans view the War on Drugs as a waste of government resources.",
"The first two drafts of the United States Declaration of Independence were written on hemp paper.Before its ban, hemp was a common cash crop for family farms in early America.",
"About one-fourth of America’s population has tried marijuana at least once. The rest are lying.",
"A recent study proved that cannabis aids both lung and breast cancer-and does not cause any form of cancer. It only cures.",
"Studies have proven that the high from cannabis is very comparable to that of a runners high.",
"The first recorded use of marijuana as a medicinal drug occurred in 2737 B.C. by Chinese emperor Shen Nung. The emperor lauded the drug as a miracle cure for rheumatism and gout.",
"There are over 200 slang terms for marijuana in the popular vernacular. Illegal Toxin is not one of them.",
"Just under 40% of high school students in the U.S. report using marijuana at least once in their life, and 20% report using it regularly.",
"The sale of bongs and paraphernalia is now illegal in Florida. The same state in which self defense with a firearm is legal.",
"Weed is the most frequently used illegal drug in the United States. And the world.",
"The growth, sale and consumption of cannabis is not regulated by the North Korean government or classified as a drug.",
"Marijuana is used to fight other drug addictions-most recently, in Bogata Colombia.",
"Accident studies have found that drivers who test positive for marijuana often show no signs of impairment, and in some instances even may drive more safely.",
"Marijuana is not physically addictive.",
"Over 50 million Americans smoke cannabis on a regular basis-while 1 in 10 teens smoke it more than 20 times a month. ",
"Health studies have proven there are no long-term health detriments from marijuana use-only benefits. And a little short term memory loss.",
"Cigarettes Kill 443,000 people a year-in the United States Alone. Vending machines kill 10 people a year. That’s 443,010 more than cannabis has ever killed.",
"The NBA does not test athletes for marijuana use in the offseason.",
"An estimated amount of over 800,000 people are arrested for marijuana each year, most of which are for simple possession.",
"An NFL lineman estimated that 50% of the league smokes marijuana on a regular basis.",
" Despite these facts, Marijuana remains federally illegal. BUT NOT FOR LONG…"];